--- sxt/multiexp/bucket_method/bucket_combination.cc
@@ -1,2 +1,0 @@
-#include "sxt/multiexp/bucket_method/bucket_combination.h"
-
